Fee Status Done
